Add a random memory tester.
The tester allocates a scratchpad and a block RAM.  It sends the same operations
to both and compares the results.  The scratchpad address space may be larger
than the block RAM space.  A mapping function translates addresses.  Using
the mapping function you can use sparse addresses on the scratchpad but touch
many pages.  I used this to test the new TLB code in leap-platforms-intel.